added github action workflow

+++ bootstrap-cluster.sh
@@ -15,10 +15,16 @@ else
 fi
 
 if [[ -n "$SECRET_VALUE" ]]; then
-    echo "$SECRET_VALUE" > "$HOME/.ssh/$KEY_NAME"
-    KEY="$HOME/.ssh/$KEY_NAME"
-    echo "running chmod 400"
-    chmod 400 "$KEY"
+    key_name="$HOME/.ssh/$KEY_NAME"
+    if [ ! -f "$key_name" ]; then
+        echo "$SECRET_VALUE" > "$HOME/.ssh/$KEY_NAME"
+        KEY="$HOME/.ssh/$KEY_NAME"
+        echo "running chmod 400"
+        chmod 400 "$KEY"
+        echo "$key_name created successfully"
+    else
+        echo "$key_name already exists."
+    fi
 else
     echo "Failed to retrieve secret value. Falling back to default key."
 fi
